move artifacts folder under /src
