+++ src/artifacts/templates/ERC20Template.sol/ERC20Template.json
