+++ src/artifacts/templates/ERC721Template.sol/ERC721Template.json
