+++ src/artifacts/utils/mock/MockERC20Decimals.sol/MockERC20Decimals.json
