+++ src/tokens/Datatoken.ts
@@ -3,8 +3,8 @@ import { AbiItem } from 'web3-utils'
 import { TransactionReceipt } from 'web3-eth'
 import { Contract } from 'web3-eth-contract'
 import Decimal from 'decimal.js'
-import defaultDatatokensAbi from '../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
-import defaultDatatokensEnterpriseAbi from '../../artifacts/templates/ERC20TemplateEnterprise.sol/ERC20TemplateEnterprise.json'
+import defaultDatatokensAbi from '../artifacts/templates/ERC20Template.sol/ERC20Template.json'
+import defaultDatatokensEnterpriseAbi from '../artifacts/templates/ERC20TemplateEnterprise.sol/ERC20TemplateEnterprise.json'
 import { LoggerInstance, getFairGasPrice } from '../utils'
 import { FreOrderParams, FreCreationParams } from '../interfaces'
 import { Nft } from './NFT'
