+++ src/tokens/NFT.ts
@@ -1,7 +1,7 @@
 import Web3 from 'web3'
 import { AbiItem } from 'web3-utils'
 import { TransactionReceipt } from 'web3-eth'
-import defaultNftAbi from '../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
+import defaultNftAbi from '../artifacts/templates/ERC721Template.sol/ERC721Template.json'
 import { LoggerInstance, getFairGasPrice, generateDtName } from '../utils'
 import { Contract } from 'web3-eth-contract'
 import { MetadataProof } from '../../src/@types'
