+++ test/unit/pools/Router.test.ts
@@ -2,15 +2,15 @@ import { assert, expect } from 'chai'
 import { AbiItem } from 'web3-utils/types'
 import Web3 from 'web3'
 import { TestContractHandler } from '../../TestContractHandler'
-import ERC721Factory from '../../../artifacts/ERC721Factory.sol/ERC721Factory.json'
-import ERC721Template from '../../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
-import SideStaking from '../../../artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
-import FactoryRouter from '../../../artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
-import ERC20Template from '../../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
-import Dispenser from '../../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
-import FixedRate from '../../../artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
-import OPFCommunityFeeCollector from '../../../artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
-import PoolTemplate from '../../../artifacts/pools/balancer/BPool.sol/BPool.json'
+import ERC721Factory from '../../artifacts/ERC721Factory.sol/ERC721Factory.json'
+import ERC721Template from '../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
+import SideStaking from '../../artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
+import FactoryRouter from '../../artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
+import ERC20Template from '../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
+import Dispenser from '../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
+import FixedRate from '../../artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
+import OPFCommunityFeeCollector from '../../artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
+import PoolTemplate from '../../artifacts/pools/balancer/BPool.sol/BPool.json'
 import { NftFactory, NftCreateData } from '../../../src/factories/NFTFactory'
 import { Router } from '../../../src/pools/Router'
 import { Erc20CreateParams, PoolCreationParams, Operation } from '../../../src/interfaces'
