+++ test/unit/pools/ssContracts/SideStaking.test.ts
@@ -3,15 +3,15 @@ import { AbiItem } from 'web3-utils/types'
 import { TestContractHandler } from '../../../TestContractHandler'
 import { Contract } from 'web3-eth-contract'
 import Web3 from 'web3'
-import ERC721Factory from '../../../../artifacts/ERC721Factory.sol/ERC721Factory.json'
-import ERC721Template from '../../../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
-import SSContract from '../../../../artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
-import FactoryRouter from '../../../../artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
-import ERC20Template from '../../../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
-import Dispenser from '../../../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
-import FixedRate from '../../../../artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
-import PoolTemplate from '../../../../artifacts/pools/balancer/BPool.sol/BPool.json'
-import OPFCollector from '../../../../artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
+import ERC721Factory from '../../../artifacts/ERC721Factory.sol/ERC721Factory.json'
+import ERC721Template from '../../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
+import SSContract from '../../../artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
+import FactoryRouter from '../../../artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
+import ERC20Template from '../../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
+import Dispenser from '../../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
+import FixedRate from '../../../artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
+import PoolTemplate from '../../../artifacts/pools/balancer/BPool.sol/BPool.json'
+import OPFCollector from '../../../artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
 import { LoggerInstance } from '../../../../src/utils'
 import { NftFactory, NftCreateData } from '../../../../src/factories/NFTFactory'
 import { Pool } from '../../../../src/pools/balancer/Pool'
